# packages/api-types/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Every non-2xx response has this shape (see errors.py).",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "RiskSettingsIO": {
        "additionalProperties": false,
        "properties": {
          "kill_switch": {
            "default": false,
            "title": "Kill Switch",
            "type": "boolean"
          },
          "max_daily_loss": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1000000000.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Daily Loss"
          },
          "max_daily_profit": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1000000000.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Daily Profit"
          },
          "max_open_positions": {
            "anyOf": [
              {
                "maximum": 1000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Open Positions"
          },
          "max_trades_per_day": {
            "anyOf": [
              {
                "maximum": 10000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Trades Per Day"
          }
        },
        "title": "RiskSettingsIO",
        "type": "object"
      },
      "RunDetail": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/RunEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "orders": {
            "items": {
              "$ref": "#/components/schemas/OrderOut"
            },
            "title": "Orders",
            "type": "array"
          },
          "positions": {
            "items": {
              "$ref": "#/components/schemas/PositionOut"
            },
            "title": "Positions",
            "type": "array"
          },
          "run": {
            "$ref": "#/components/schemas/RunOut"
          }
        },
        "required": [
          "run",
          "positions",
          "orders",
          "events"
        ],
        "title": "RunDetail",
        "type": "object"
      },
...
      "RunOut": {
        "properties": {
          "broker_account_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broker Account Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "engine_stale": {
            "title": "Engine Stale",
            "type": "boolean"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "heartbeat_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heartbeat At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "mode": {
            "enum": [
              "paper",
              "live"
            ],
            "title": "Mode",
            "type": "string"
          },
          "multiplier": {
            "title": "Multiplier",
            "type": "integer"
          },
          "open_positions": {
            "title": "Open Positions",
            "type": "integer"
          },
          "realized_pnl": {
            "title": "Realized Pnl",
            "type": "number"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "stopping",
              "stopped",
              "completed",
              "error"
            ],
            "title": "Status",
            "type": "string"
          },
          "stop_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Reason"
          },
          "stopped_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stopped At"
          },
          "strategy_id": {
            "format": "uuid",
            "title": "Strategy Id",
            "type": "string"
          },
          "strategy_name": {
            "title": "Strategy Name",
            "type": "string"
          },
          "underlying": {
            "title": "Underlying",
            "type": "string"
          },
          "unrealized_pnl": {
            "title": "Unrealized Pnl",
            "type": "number"
          }
        },
        "required": [
          "id",
          "strategy_id",
          "strategy_name",
          "kind",
          "underlying",
          "mode",
          "status",
          "multiplier",
          "broker_account_id",
          "realized_pnl",
          "unrealized_pnl",
          "open_positions",
          "created_at",
          "started_at",
          "stopped_at",
          "heartbeat_at",
          "engine_stale",
          "stop_reason",
          "error"
        ],
        "title": "RunOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AlgoEarning API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/me/risk": {
      "get": {
        "operationId": "get_risk_v1_me_risk_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskSettingsIO"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sign in required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unavailable"
          }
        },
        "summary": "Get Risk",
        "tags": [
          "runs"
        ]
      },
      "put": {
        "operationId": "put_risk_v1_me_risk_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskSettingsIO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskSettingsIO"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sign in required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unavailable"
          }
        },
        "summary": "Put Risk",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/v1/runs": {
      "get": {
        "operationId": "list_runs_v1_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "active",
            "required": false,
            "schema": {
              "default": true,
              "title": "Active",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunOut"
                  },
                  "title": "Response List Runs V1 Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sign in required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unavailable"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      }
    },
    "/v1/runs/stop-all": {
      "post": {
        "operationId": "stop_all_v1_runs_stop_all_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunOut"
                  },
                  "title": "Response Stop All V1 Runs Stop All Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sign in required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unavailable"
          }
        },
        "summary": "Stop All",
        "tags": [
          "runs"
        ]
      }
    },
    "/v1/runs/{run_id}": {
      "get": {
        "operationId": "get_run_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sign in required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unavailable"
          }
        },
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/v1/runs/{run_id}/stop": {
      "post": {
        "description": "Square off the run's open positions and stop it (the engine does it within a second or two).",
        "operationId": "stop_run_v1_runs__run_id__stop_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sign in required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unavailable"
          }
        },
        "summary": "Stop Run",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/v1/strategies/{strategy_id}/deploy": {
      "post": {
        "operationId": "deploy_v1_strategies__strategy_id__deploy_post",
        "parameters": [
          {
            "in": "path",
            "name": "strategy_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Strategy Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeployIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sign in required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unavailable"
          }
        },
        "summary": "Deploy",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}